Fix issues revealed by -Wextra flag.
Fix assorted warnings raised by the -Wextra flag, including
a few significant items. Reduced local build warnings from
over 9200 warnings to about 650.

At least one unicast send routine would drop all errors due
to type issues.

Reviewed-on: https://gerrit.iotivity.org/gerrit/2018
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>

+++ resource/csdk/connectivity/src/ip_adapter/linux/caipnwmonitor.c
@@ -75,7 +75,9 @@ u_arraylist_t *CAIPGetInterfaceInformation(int desiredIndex)
         {
             CAInterface_t *ifitem = (CAInterface_t *)u_arraylist_get(iflist, i);
 
-            if (ifitem && ifitem->index == ifindex && ifitem->family == family)
+            if (ifitem
+                && (int)ifitem->index == ifindex
+                && ifitem->family == (uint16_t)family)
             {
                 already = true;
                 break;
